feat: Change types

+++ common/src/constants.rs
@@ -1,11 +1,11 @@
 /// Maximum number of identical tiles
-pub const MAX_NUM_TILE: i8 = 4;
+pub const MAX_NUM_TILE: usize = 4;
 
 /// Number of tile types
-pub const NUM_TILE_TYPE: i8 = 34;
+pub const NUM_TILE_TYPE: usize = 34;
 
 /// Minimum number of tiles in a hand
-pub const MIN_HAND_SIZE: i8 = 1;
+pub const MIN_HAND_SIZE: usize = 1;
 
 /// Maximum number of tiles in a hand
-pub const MAX_HAND_SIZE: i8 = 14;
+pub const MAX_HAND_SIZE: usize = 14;
